Resource read and write from faunadb
Removed an unnecessary AddResource function, as all get and post requests go through the allResources function.
Form.js -
Updated a few form attributes to match the name of the schema in faunadb. Added a timestamp attribute to send over, and updated the function endpoint for the add resource post call.

index.js -
changed the jsx resource render logic to look to the key of "topic" instead of "language/topic", reflecting an update to the resource schema. Same value, updated key.